Lecture 34 Movement State P2

+++ Udemy 2d Platformer/Assets/_Scripts/States/MovementData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class MovementData : MonoBehaviour
 {
     internal int horizontalMovementDirection;
-    internal int currentSpeed;
+    internal float currentSpeed;
     internal Vector2 currentVelocity;
 
     // Start is called before the first frame update
